requirements.txt
wikipedia
mlflow==2.14.1
pdfplumber
pypdf
langchain==0.3.10
langchain-community==0.3.10
openai
langchain-openai==0.2.11
langchainhub
langgraph==0.2.56
python-dotenv~=1.0.1
pydantic~=2.10.2
tavily-python
faiss-cpu
langchain-anthropic~=0.3.0
langchain-experimental~=0.3.3
rank_bm25
rank-llm==0.12.8
chromadb
langchain_cohere
PyYAML==6.0.2
docling~=2.15.1
langchain_text_splitters
langchain-core~=0.3.22
typing_extensions~=4.12.2
langchain-text-splitters~=0.3.2
rank-bm25~=0.2.2
youtube-search==2.1.2
youtube-transcript-api==0.6.3
pymongo==4.11.1
chromadb==0.6.3
chromadb-client==0.6.3
faiss-cpu==1.10.0
langchain-openai